Extract IngestJobStatusReporterTestData from IngestJobStatusReporterTestHelper
