+++ java/clients/src/test/java/sleeper/status/report/ingest/job/StandardIngestJobStatusReporterDetailedQueryTest.java
@@ -25,10 +25,10 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.ClientTestUtils.example;
 import static sleeper.status.report.ingest.job.IngestJobStatusReporter.QueryType;
+import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestData.jobWithMultipleRuns;
+import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestData.jobsWithLargeAndDecimalStatistics;
+import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestData.mixedJobStatuses;
 import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.getStandardReport;
-import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.jobWithMultipleRuns;
-import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.jobsWithLargeAndDecimalStatistics;
-import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.mixedJobStatuses;
 import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.replaceBracketedJobIds;
 
 public class StandardIngestJobStatusReporterDetailedQueryTest {
