+++ java/clients/src/test/java/sleeper/status/report/ingest/job/StandardIngestJobStatusReporterUnfinishedQueryTest.java
@@ -25,8 +25,8 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.ClientTestUtils.example;
 import static sleeper.status.report.ingest.job.IngestJobStatusReporter.QueryType;
+import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestData.mixedUnfinishedJobStatuses;
 import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.getStandardReport;
-import static sleeper.status.report.ingest.job.IngestJobStatusReporterTestHelper.mixedUnfinishedJobStatuses;
 
 public class StandardIngestJobStatusReporterUnfinishedQueryTest {
     @Test
